+++ pyobjc-core/t.m
@@ -1,6 +1,0 @@
-#import <Foundation/Foundation.h>
-
-int main(void)
-{
-	NSLog(@"%llu", [[NSNumber numberWithDouble:-127.6] unsignedLongLongValue]);
-}